requirements.txt
dropbox==12.0.2
python-vlc==3.0.21203